Allow chef-solo for agent recipe
The `ambari::agent` recipe should allow the user to specify the
Ambari server and the recipe should respect this setting. In the
case of Chef Solo, there could also be the `chef-solo-search`
cookbook in the run_list, which will give Chef Solo searching
capabilities.

Also, updated search to use recipes over run_list, so it picks up
use of `ambari::server` via a wrapper cookbook or role.