Added mechanics to allow different db types
The changes involves using attributes to allow for a one-step
installation of the configured database backend, currently only postgres
and mysql is supported for ubuntu/debian only, but more can be added.

Also removed the installation of postgres in the cookbook since
ambari-server package depends on it and will install/configure it
appropriately.

All of these changes was initially written by @Raboo some time ago. I
just merged changes to @jp's master.