added a block that removes 127.0.1.1 from hosts
Debian and ubuntu distros adds a entry in /etc/hosts that points the
hostname to 127.0.0.1, this won't work properly with ambari or hadoop so
we remove it here, even though this might be better off in a
wrapper-cookbook.

All of these changes was initially written by @Raboo some time ago. I
just merged changes to @jp's master.